New Interfaces struct emiting interface related code
